Update app.R
Made all the necessary text bold. Added background image, and had to add in topmost layer of FluidPage in order to allow for background image. Added a last tabpanel because I couldn't find any other way to add in the logo to the NavBar, it's a small bug because you can click on the empty tab and it'll be a blank tab, but not sure if it's big enough to need fixing.